+++ Netology_IB_Instruments.xcodeproj/xcuserdata/aleksandrpozdnikin.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,17 +7,17 @@
 		<key>Netology_IB_Instruments.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>2</integer>
+			<integer>0</integer>
 		</dict>
 		<key>Netology_IB_Instruments_release1.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>3</integer>
+			<integer>1</integer>
 		</dict>
 		<key>StorageService.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>4</integer>
+			<integer>5</integer>
 		</dict>
 	</dict>
 	<key>SuppressBuildableAutocreation</key>
