+++ Netology_IB_Instruments.xcworkspace/xcuserdata/aleksandrpozdnikin.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Bucket
-   uuid = "273AA00D-0592-4528-A714-9C6C3626F9BA"
+   uuid = "7AFC811A-AACE-4BA7-9081-A0F59DC48950"
    type = "0"
    version = "2.0">
 </Bucket>
