+++ Netology_IB_Instruments/AppDelegate.swift
@@ -32,5 +32,61 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         window?.rootViewController = tabBarController
         return true
     }
+    
+    func applicationWillResignActive(_ application: UIApplication) {
+
+        // Sent when the application is about to move from active to inactive state. This can occur for certain types of temporary interruptions (such as an incoming phone call or SMS message) or when the user quits the application and it begins the transition to the background state.
+
+        // Use this method to pause ongoing tasks, disable timers, and invalidate graphics rendering callbacks. Games should use this method to pause the game.
+
+        print(#function)
+
+    }
+
+
+
+
+    func applicationDidEnterBackground(_ application: UIApplication) {
+
+        // Use this method to release shared resources, save user data, invalidate timers, and store enough application state information to restore your application to its current state in case it is terminated later.
+
+        // If your application supports background execution, this method is called instead of applicationWillTerminate: when the user quits.
+
+        print(#function)
+
+    }
+
+
+
+
+    func applicationWillEnterForeground(_ application: UIApplication) {
+
+        // Called as part of the transition from the background to the active state; here you can undo many of the changes made on entering the background.
+
+        print(#function)
+
+    }
+
+
+
+
+    func applicationDidBecomeActive(_ application: UIApplication) {
+
+        // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
+
+        print(#function)
+
+    }
+
+
+
+
+    func applicationWillTerminate(_ application: UIApplication) {
+
+        // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
+
+        print(#function)
+
+    }
 }
